refactor(questions): Remove unused useEffect and clean up questions initialization
